requirements.txt
asgiref==3.2.3
bottle==0.12.18
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0
future==0.18.2
gspread==3.1.0
gspread-formatting==0.3.7
httplib2==0.14.0
idna==2.8
numpy==1.18.2
oauth2client==4.1.3
pandas==1.0.3
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pymongo==3.10.0
python-dateutil==2.8.1
python-telegram-bot==12.2.0
pytz==2019.3
query-string==2019.4.13
requests==2.22.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.7
